Reorganize for clarity
