+++ hunch/NotionItems/Color.swift
@@ -1,0 +1,30 @@
+//
+//  Color.swift
+//  hunch
+//
+//  Created by Adam Wulf on 11/12/24.
+//
+
+import Foundation
+
+enum Color: String, Codable {
+    case plain = "default"
+    case gray
+    case brown
+    case orange
+    case yellow
+    case green
+    case blue
+    case purple
+    case pink
+    case red
+    case grayBackground = "gray_background"
+    case brownBackground = "brown_background"
+    case orangeBackground = "orange_background"
+    case yellowBackground = "yellow_background"
+    case greenBackground = "green_background"
+    case blueBackground = "blue_background"
+    case purpleBackground = "purple_background"
+    case pinkBackground = "pink_background"
+    case redBackground = "red_background"
+}
