+++ hunch/NotionItems/NotionItem.swift
@@ -1,0 +1,14 @@
+//
+//  NotionItem.swift
+//  hunch
+//
+//  Created by Adam Wulf on 11/12/24.
+//
+
+import Foundation
+
+protocol NotionItem: Codable, CustomStringConvertible {
+    var object: String { get }
+    var id: String { get }
+    var parent: Parent? { get }
+}
